Implement column toggle button support for dynatable
And as a test case, implement the toggle button for the drafts search
result page.
